pragya_ai/requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
ccxt==4.4.86
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dateparser==1.2.1
fonttools==4.59.0
frozenlist==1.6.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
MetaTrader5==5.0.5120
multidict==6.4.4
numpy==1.24.4
packaging==25.0
pandas==2.2.3
pillow==11.3.0
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
pycryptodome==3.23.0
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.7
ta==0.11.0
TA-Lib @ file:///D:/Python/TA_Lib-0.4.0-cp310-cp310-win_amd64.whl#sha256=c4ba337d31a2c83e71696dad7ea9b2e3f0bbe9c303c06bad2de0736243702d5e
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
